+++ src/main/java/org/cloud/sonic/agent/models/HandleDes.java
@@ -1,4 +1,4 @@
-package org.cloud.sonic.agent.automation;
+package org.cloud.sonic.agent.models;
 
 /**
  * @author ZhouYiXun
