+++ src/main/java/org/cloud/sonic/agent/tests/handlers/ElseHandler.java
@@ -17,7 +17,7 @@
 package org.cloud.sonic.agent.tests.handlers;
 
 import com.alibaba.fastjson.JSONObject;
-import org.cloud.sonic.agent.automation.HandleDes;
+import org.cloud.sonic.agent.models.HandleDes;
 import org.cloud.sonic.agent.common.interfaces.StepType;
 import org.cloud.sonic.agent.enums.ConditionEnum;
 import org.cloud.sonic.agent.tests.common.RunStepThread;
