+++ src/main/java/org/cloud/sonic/agent/tools/cv/AKAZEFinder.java
@@ -22,7 +22,7 @@ import org.bytedeco.opencv.opencv_flann.Index;
 import org.bytedeco.opencv.opencv_flann.IndexParams;
 import org.bytedeco.opencv.opencv_flann.LshIndexParams;
 import org.bytedeco.opencv.opencv_flann.SearchParams;
-import org.cloud.sonic.agent.automation.FindResult;
+import org.cloud.sonic.agent.models.FindResult;
 import org.cloud.sonic.agent.tools.file.UploadTools;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
