+++ src/main/java/org/cloud/sonic/agent/tools/cv/SIFTFinder.java
@@ -19,7 +19,7 @@ package org.cloud.sonic.agent.tools.cv;
 import org.bytedeco.opencv.opencv_core.*;
 import org.bytedeco.opencv.opencv_features2d.FlannBasedMatcher;
 import org.bytedeco.opencv.opencv_xfeatures2d.SIFT;
-import org.cloud.sonic.agent.automation.FindResult;
+import org.cloud.sonic.agent.models.FindResult;
 import org.cloud.sonic.agent.tools.file.UploadTools;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
